requirements.txt
langchain_google_genai
langchain_core
streamlit
python-dotenv